+++ src/main/kotlin/dev/aaronhowser/mods/geneticsresequenced/event/player/ClickEvents.kt
@@ -1,12 +1,12 @@
 package dev.aaronhowser.mods.geneticsresequenced.event.player
 
 import dev.aaronhowser.mods.geneticsresequenced.GeneticsResequenced
+import dev.aaronhowser.mods.geneticsresequenced.datagen.tag.ModEntityTypeTagsProvider
+import dev.aaronhowser.mods.geneticsresequenced.datagen.tag.ModItemTagsProvider
 import dev.aaronhowser.mods.geneticsresequenced.entity.SupportSlime
 import dev.aaronhowser.mods.geneticsresequenced.gene.behavior.AttributeGenes
 import dev.aaronhowser.mods.geneticsresequenced.gene.behavior.ClickGenes
-import dev.aaronhowser.mods.geneticsresequenced.registry.ModItems
 import net.minecraft.world.InteractionHand
-import net.minecraft.world.entity.npc.AbstractVillager
 import net.neoforged.bus.api.SubscribeEvent
 import net.neoforged.fml.common.EventBusSubscriber
 import net.neoforged.neoforge.event.entity.EntityJoinLevelEvent
@@ -30,16 +30,14 @@ object ClickEvents {
 
     private fun checkShouldCancel(event: PlayerInteractEvent.EntityInteract) {
         val entity = event.target
-        if (entity !is AbstractVillager) return
+        if (!entity.type.`is`(ModEntityTypeTagsProvider.ALLOWS_PREVENTING_INTERACTION)) return
 
         val mainHandStack = event.entity.getItemInHand(InteractionHand.MAIN_HAND)
         val offHandStack = event.entity.getItemInHand(InteractionHand.OFF_HAND)
 
         if (
-            mainHandStack.item == ModItems.METAL_SYRINGE.get()
-            || offHandStack.item == ModItems.METAL_SYRINGE.get()
-            || mainHandStack.item == ModItems.SCRAPER.get()
-            || offHandStack.item == ModItems.SCRAPER.get()
+            mainHandStack.`is`(ModItemTagsProvider.PREVENTS_SOME_MOB_INTERACTION)
+            || offHandStack.`is`(ModItemTagsProvider.PREVENTS_SOME_MOB_INTERACTION)
         ) {
             event.isCanceled = true
         }